requirements.txt
# Core
pyyaml>=6.0
customtkinter>=5.2.0
requests>=2.28.0
beautifulsoup4>=4.11.0
colorama>=0.4.4
click>=8.0.0
tqdm>=4.64.0
jinja2>=3.1.0

# System Scanning
bandit>=1.7.0
safety>=2.0.0

# Network Scanning
python-nmap>=0.7.1
scapy>=2.4.5
cryptography>=37.0.0
paramiko>=2.11.0

# Android Scanning
androguard>=3.4.0
apkfile>=0.1.7

# Web Scanning
selenium>=4.0.0
urllib3>=1.26.0
aiohttp>=3.9.0
pybuster>=0.2.1
pydirbuster>=0.5

# UI
PySide6>=6.6.0
qtawesome>=1.3.0
# Additional Tools
shodan>=1.28.0
pyshark>=0.6.0
pwntools>=4.13.0
unicorn>=2.0.1
capstone>=5.0.1
ROPGadget>=7.4
pyserial>=3.5
rpyc>=5.3.1
tldextract>=5.1.1
pyelftools>=0.30
psutil>=5.9.6
alembic>=1.13.1
dataset>=1.6.2
ipython>=8.18.1

# Bluetooth
bleak>=0.21.1
pypiwin32>=223

# Intruder Pack - CrackMapExec Dependencies
impacket>=0.11.0
pypykatz>=0.6.0
rich>=13.0.0
netaddr>=0.8.0
pyOpenSSL>=23.0.0
pycryptodome>=3.17.0
termcolor>=2.2.0
terminaltables>=3.1.0
aioconsole>=0.6.0
xmltodict>=0.13.0

# Intruder Pack - XSStrike Dependencies
# XSStrike is a subprocess-invoked tool (not a pip package).
# Clone: git clone https://github.com/s0md3v/XSStrike engines/intruder/xsstrike_src
# These are XSStrike's own runtime deps (from its requirements.txt):
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.12.0
prettytable>=3.9.0
tld>=0.12.7

# Intruder Pack - HTTPie Dependencies
# httpie is a proper pip library — its internal Python API is used directly
httpie>=3.2.0
# Intruder Pack - Photon OSINT Crawler Dependencies
# Photon engine is a native implementation — no external photon package needed.
# These are its runtime deps (most already present above):
tldextract>=5.1.1
dnspython>=2.4.0
# Intruder Pack - Wapiti Engine Dependencies
wapiti3>=3.1.6
